Split LTX2.3 pipeline tests by feature
